ci: rename the osx ens domains file
