+++ configs/src/deployments/index.ts
@@ -1,4 +1,4 @@
 export * from './types';
 export * from './getters';
 export * from './contracts';
-export * from './osx-ens-domains';
+export * from './ens';
